handled exception

+++ src/main/java/com/monk/ecom/service/impl/UpdateCouponServiceImpl.java
@@ -12,6 +12,7 @@ import com.monk.ecom.domain.Coupon;
 import com.monk.ecom.domain.CreationCriteria;
 import com.monk.ecom.domain.Product;
 import com.monk.ecom.domain.ProductDetails;
+import com.monk.ecom.exception.CouponNotFoundException;
 import com.monk.ecom.repository.BuyAndGetCouponRepository;
 import com.monk.ecom.repository.CouponForCartRepository;
 import com.monk.ecom.repository.CouponRepository;
@@ -24,7 +25,7 @@ import com.monk.ecom.util.couponUtility;
 public class UpdateCouponServiceImpl implements UpdateCouponService {
 
     @Autowired
-    private CouponRepository repository;
+    private CouponRepository couponRepository;
 
     @Autowired
     private ProductRepository productRepo;
@@ -42,42 +43,47 @@ public class UpdateCouponServiceImpl implements UpdateCouponService {
     private couponUtility utility;
 
     @Override
-    public void updateCoupon(long id, CreationCriteria couponCriteria) {
-        Coupon coupon = repository.findById(id).get();
-        ProductDetails details = couponCriteria.getDetails();
-        if (coupon.getType().equalsIgnoreCase("cart-wise")) {
-            String description = utility.getCartWiseDiscountDescription(details);
-            repository.updateCouponDiscountAndDescription(id, details.getDiscount(), description);
+    public void updateCoupon(long id, CreationCriteria couponCriteria) throws CouponNotFoundException {
+        Optional<Coupon> couponOpt = couponRepository.findById(id);
+        if (couponOpt.isPresent()) {
+            Coupon coupon = couponOpt.get();
+            ProductDetails details = couponCriteria.getDetails();
+            if (coupon.getType().equalsIgnoreCase("cart-wise")) {
+                String description = utility.getCartWiseDiscountDescription(details);
+                couponRepository.updateCouponDiscountAndDescription(id, details.getDiscount(), description);
 
-            couponForCartRepo.updateCouponForCart(id, details.getThreshold(), details.getDiscount(),
-                    description);
-        } else if (coupon.getType().equalsIgnoreCase("product-wise")) {
-            String description = utility.getProductWiseDiscountDescription(details);
-            repository.updateCouponDiscountAndDescription(id, details.getDiscount(), description);
-            Product product = productRepo.findById(details.getProductId()).get();
-            productCouponMapRepo.updateProductCouponMap(id, details.getProductId(),
-                    product.getPrice(), details.getDiscount(), product.getPrice() - details.getDiscount());
-        } else if (coupon.getType().equalsIgnoreCase("BxGy")) {
-            String description = utility.getBxGyDiscountDescription(details);
-            repository.updateCouponDiscountAndDescription(id, details.getDiscount(), description);
-            buyAndGetCouponRepo.deleteByCouponId(id);
-            List<BuyAndGetCouponMap> entiityUpdateList = new ArrayList<>();
-            for (Product buyProd : details.getBuyProducts()) {
-                for (Product getProd : details.getGetProducts()) {
-                    Product product = productRepo.findById(getProd.getProductId()).get();
-                    float discount = product.getPrice() * getProd.getQuantity();
-                    BuyAndGetCouponMap bngCouponMap = new BuyAndGetCouponMap(id, buyProd.getProductId(),
-                            buyProd.getQuantity(), getProd.getProductId(), getProd.getQuantity(), discount);
-                    entiityUpdateList.add(bngCouponMap);
+                couponForCartRepo.updateCouponForCart(id, details.getThreshold(), details.getDiscount(),
+                        description);
+            } else if (coupon.getType().equalsIgnoreCase("product-wise")) {
+                String description = utility.getProductWiseDiscountDescription(details);
+                couponRepository.updateCouponDiscountAndDescription(id, details.getDiscount(), description);
+                Product product = productRepo.findById(details.getProductId()).get();
+                productCouponMapRepo.updateProductCouponMap(id, details.getProductId(),
+                        product.getPrice(), details.getDiscount(), product.getPrice() - details.getDiscount());
+            } else if (coupon.getType().equalsIgnoreCase("BxGy")) {
+                String description = utility.getBxGyDiscountDescription(details);
+                couponRepository.updateCouponDiscountAndDescription(id, details.getDiscount(), description);
+                buyAndGetCouponRepo.deleteByCouponId(id);
+                List<BuyAndGetCouponMap> entiityUpdateList = new ArrayList<>();
+                for (Product buyProd : details.getBuyProducts()) {
+                    for (Product getProd : details.getGetProducts()) {
+                        Product product = productRepo.findById(getProd.getProductId()).get();
+                        float discount = product.getPrice() * getProd.getQuantity();
+                        BuyAndGetCouponMap bngCouponMap = new BuyAndGetCouponMap(id, buyProd.getProductId(),
+                                buyProd.getQuantity(), getProd.getProductId(), getProd.getQuantity(), discount);
+                        entiityUpdateList.add(bngCouponMap);
+                    }
                 }
+                buyAndGetCouponRepo.saveAll(entiityUpdateList);
+            } else {
+                throw new CouponNotFoundException("Coupon with couponId: " + id + " does not exist");
             }
-            buyAndGetCouponRepo.saveAll(entiityUpdateList);
         }
     }
 
     @Override
-    public void deleteCoupon(long id) {
-        Optional<Coupon> OptionalCoupon = repository.findById(id);
+    public void deleteCoupon(long id) throws CouponNotFoundException {
+        Optional<Coupon> OptionalCoupon = couponRepository.findById(id);
         if (OptionalCoupon.isPresent()) {
             Coupon coupon = OptionalCoupon.get();
             if (coupon.getType().equalsIgnoreCase("cart-wise")) {
@@ -87,9 +93,9 @@ public class UpdateCouponServiceImpl implements UpdateCouponService {
             } else if (coupon.getType().equalsIgnoreCase("BxGy")) {
                 buyAndGetCouponRepo.deleteByCouponId(id);
             }
-            repository.deleteById(id);
+            couponRepository.deleteById(id);
         } else {
-            // coupon not present error;
+            throw new CouponNotFoundException("Coupon with couponId: " + id + " does not exist");
         }
     }
